% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelMiss.R
\name{DelMiss}
\alias{DelMiss}
\title{Delete variables with missing values}
\usage{
DelMiss(eSet = eSet)
}
\arguments{
\item{eSet}{An R6 class object.}
}
\value{
An R6 class object containing the variable(s) without missing values.
}
\description{
Whether to delete missing variables with low variance. The default option is "yes".
If skipped, it may result in failure during modeling.
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
}
\author{
Bin Wang
}
