% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exit.R
\name{Exit}
\alias{Exit}
\title{End the module analysis}
\usage{
Exit(eSet = eSet)
}
\arguments{
\item{eSet}{An R6 class object}
}
\description{
End the module analysis
}
\examples{
   eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
eSet = TransType(eSet = eSet,Vars = c("Y1", "C1","E203","E204","E207","E209"),TypeTo = "factor")
eSet = TransScale(eSet = eSet,Group = T,Vars = "all.e",Method = "normal",
                  Direct = "positive",RangeLow = 0,RangeUpper = 1)
eSet = TransDistr(eSet = eSet,Vars = c("C2", "C3"), Method = "log2")
eSet = TransDummy(eSet = eSet,Vars = "default")
eSet = MulGroupsCros(eSet = eSet,Groups = c("Exposure","Biomarker"),
                    VarsY = c("Y1"),VarsC = "all.c",TuneMethod = "grid_search",TuneNum = 5,RsmpMethod = "cv",
                    Folds = 5,VarsImpThr = 0.85,SG_Lrns = c("rf","xgboost"))
eSet = VizMulGroupsCros(eSet = eSet,VarsY = c("Y1"),NodeNum = 100,EdgeThr = 0.5,
                       Layout = "force-directed",Brightness = "light",Palette = "default1")
Exit(eSet = eSet)
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
