% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitEX.R
\name{InitEX}
\alias{InitEX}
\title{Initialize Expred module}
\usage{
InitEX()
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitEX.}

\item{FileDirIn}{"default"}

\item{FileDirOut}{"default"}
}
\value{
An R6 class object.
}
\description{
Initialize Expred analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
Expred module is designed to integrate the multi-group data to predict the incidence risk.
It mainly aims to construct various stacked generalization(SG) models to predict the probability
of outcome incidence, as well as providing the statistical explanation. In addition, the module can provide
visualization plots with high quality of the final calculation results to make it easier for users to understand.
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
