% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadEX.R
\name{LoadEX}
\alias{LoadEX}
\title{Load data file for multigroups module}
\usage{
LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
}
\arguments{
\item{eSet}{An R6 class object}

\item{UseExample}{chr. Method of uploading data. If "default", user should upload their own data files,
or use "example#1" provided by this module.}

\item{FileDirExpo}{chr. Input file directory, e.g. "D:/test/eg_data.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{FileDirVoca}{chr. Input file vocabulary, e.g. "D:/test/eg_voca.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Upload data file for multigroups module.
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
