% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransDistr.R
\name{TransDistr}
\alias{TransDistr}
\title{Transform variable distribution}
\usage{
TransDistr(eSet = eSet,Vars = c("C2", "C3"), Method = "log2")
}
\arguments{
\item{eSet}{An R6 class object}

\item{Vars}{Variables to be imputed. Available options include:
"all.e", all exposure variables; "all.c", all covariates;
"all.ce", combination of All E and All C. Users can also choose available variables,e.g."E1,E2,E3".}

\item{Method}{chr. Methods used for imputation. Available options include "lod" or "cart".
For "lod" method, limit of detection (LOD) should be included in the "Vocabulary" file.}
}
\value{
An R6 class object containing the variable(s) after transforming distribution.
}
\description{
Transform variable distribution
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
eSet = TransType(eSet = eSet,Vars = c("Y1", "C1","E203","E204","E207","E209"),TypeTo = "factor")
eSet = TransScale(eSet = eSet,Group = T,Vars = "all.e",Method = "normal",
                  Direct = "positive",RangeLow = 0,RangeUpper = 1)
eSet = TransDistr(eSet = eSet,Vars = c("C2", "C3"), Method = "log2")
}
\author{
Bin Wang
}
