% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransImput.R
\name{TransImput}
\alias{TransImput}
\title{Missing data imputation.}
\usage{
TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
}
\arguments{
\item{eSet}{An R6 class object.}

\item{Group}{lgl. Whether to separate dataset into train and test data for processing data.}

\item{Vars}{Variables to be imputed. Available options include:
"all.e", all exposure variables; "all.c", all covariates;
"all.ce", combination of All E and All C. Users can also choose available variables,e.g."E1,E2,E3".}

\item{Method}{Methods used for imputation. Available options include "lod" or "cart" methods.
For "lod" method, limit of detection (LOD) should be included in the "Vocabulary" file.}
}
\value{
An R6 class object containing variable(s) with imputation.
}
\description{
Missing data imputation.
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSetUseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
}
\author{
Bin Wang
}
