% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransScale.R
\name{TransScale}
\alias{TransScale}
\title{Scale variables}
\usage{
TransScale(eSet = eSet,Group = T,Vars = "all.e",Method = "normal",
                  Direct = "positive",RangeLow = 0,RangeUpper = 1)
}
\arguments{
\item{eSet}{An R6 class object.}

\item{Group}{lgl. T (or TRUE) and F (or FALSE). Whether to separate dataset into train and test data for processing data.}

\item{Vars}{Variables to be imputed. Available options include:
"all.e", all exposure variables; "all.c", all covariates;
"all.ce", combination of All E and All C. Users can also choose available variables,e.g."E1,E2,E3".}

\item{Method}{chr. Scaling methods. Available options include "normal" and "range".}

\item{Direct}{chr. Direction to be transformed, Available options include "positive" and "negative".}

\item{RangeLow}{num. Lower limit for range method.}

\item{RangeUpper}{num. Upper limit for range method. It should be greater than the lower limit.}
}
\value{
An R6 class object containing the variable(s) after scaling data.
}
\description{
Scale variables
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
eSet = TransType(eSet = eSet,Vars = c("Y1", "C1","E203","E204","E207","E209"),TypeTo = "factor")
eSet = TransScale(eSet = eSet,Group = T,Vars = "all.e",Method = "normal",
                  Direct = "positive",RangeLow = 0,RangeUpper = 1)
}
\author{
Bin Wang
}
