% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransType.R
\name{TransType}
\alias{TransType}
\title{Transform data type}
\usage{
TransType(eSet = eSet,Vars,TypeTo)
}
\arguments{
\item{eSet}{An R6 class object}

\item{Vars}{Variables to be imputed. Available options include:
"all.e", all exposure variables; "all.c", all covariates;
"all.ce", combination of All E and All C. Users can also choose available variables,e.g."E1,E2,E3".}

\item{TypeTo}{chr. Indicate the type of the chosen variables to be transformed into.
Available options include "integer", "numeric", "character", "factor", "logical", and "date".}
}
\value{
An R6 class object containing the variable(s) after transforming data type.
}
\description{
Transform data type
}
\examples{
eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
eSet = TransType(eSet = eSet,Vars = c("Y1", "C1","E203","E204","E207","E209"),TypeTo = "factor")
}
\author{
Bin Wang
}
