% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizMulGroupsCros.R
\name{VizMulGroupsCros}
\alias{VizMulGroupsCros}
\title{Visualize multigroups model results}
\usage{
eSet = VizMulGroupsCros(eSet = eSet,VarsY = c("Y1"),NodeNum = 100,EdgeThr = 0.5,
                       Layout = "force-directed",Brightness = "light",Palette = "default1")
}
\arguments{
\item{eSet}{An R6 class object.}

\item{VarsY}{chr. Outcome variable for modeling. Only one variable can be entered.}

\item{NodeNum}{num. Number of nodes in the network plot. The maximum number is generated by multigroups function.
User can set a smaller value if needed.}

\item{EdgeThr}{num. Threshold of correlation coefficient ranging 0-1 for generating the concerned edges of the network plot.}

\item{Mode}{ggraph}

\item{Layout}{chr. Visualization layout. Available options include "force-directed" and "degree-circle".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2" and other options about some
journal preference styles including "cell", "nature", "science", "lancet", "nejm", etc.}
}
\value{
An R6 class object containing seven elements. The elements of that object include:
(1) "Importance_plot": Plots for the importance of features after modeling a single group from different learners.
(2) "Measures_boxplot": Plots for the r-square value of the single group model built by different learners.
(3) "NetWork_State": A list containing dataframes that contain nodes and edges used to draw
intergroup network from models built by different learners.
(4) "Nodeplot": Intergroup node plots from models built by different models.
(5) "Networkplot": Intergroup network plots from models built by different models.
(6) "Prediction_train_plot": Plots for the prediction value of the train set in the SG model built by different combinations of learners.
(7) "Prediction_test_plot": Plots for the prediction value of the test set in the SG model built by different combinations of learners.
}
\description{
VizMulGroupsCros function is mainly aimed to visualize the modeling results calculated by MulGroupsCros function.
It can provide plots with high quality of the final results to make it easier for users to understand.
}
\details{
You can get different styles of images by selecting different parameters.
}
\examples{
   eSet = InitEX(PID = "EX", FileDirIn = "default", FileDirOut = "default")
eSet = LoadEX(eSet = eSet,UseExample = "default",FileDirExpo = "examdata.xlsx",FileDirVoca = "examvoca.xlsx")
eSet = TransImput(eSet = eSet,Group = T,Vars = "all.e",Method = "lod")
eSet = DelLowVar(eSet = eSet)
eSet = DelMiss(eSet = eSet)
eSet = TransType(eSet = eSet,Vars = c("Y1", "C1","E203","E204","E207","E209"),TypeTo = "factor")
eSet = TransScale(eSet = eSet,Group = T,Vars = "all.e",Method = "normal",
                  Direct = "positive",RangeLow = 0,RangeUpper = 1)
eSet = TransDistr(eSet = eSet,Vars = c("C2", "C3"), Method = "log2")
eSet = TransDummy(eSet = eSet,Vars = "default")
eSet = MulGroupsCros(eSet = eSet,Groups = c("Exposure","Biomarker"),
                    VarsY = c("Y1"),VarsC = "all.c",TuneMethod = "grid_search",TuneNum = 5,RsmpMethod = "cv",
                    Folds = 5,VarsImpThr = 0.85,SG_Lrns = c("rf","xgboost"))
eSet = VizMulGroupsCros(eSet = eSet,VarsY = c("Y1"),NodeNum = 100,EdgeThr = 0.5,
                       Layout = "force-directed",Brightness = "light",Palette = "default1")
}
\author{
Guohuan Zhang, Yuting Wang, Ning Gao, Bin Wang (corresponding author)
}
